Support load plugin and config rng

+++ lume/LumeMeta/src/call_context.cpp
@@ -1,18 +1,3 @@
-/*
- * Copyright (c) 2024 Huawei Device Co., Ltd.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 
 #include "call_context.h"
 
@@ -52,7 +37,7 @@ DefaultCallContext& DefaultCallContext::operator=(const DefaultCallContext& othe
     }
     params_.clear();
     params_.resize(other.params_.size());
-    for (size_t i = 0; i != params_.size(); ++i) {
+    for (int i = 0; i != params_.size(); ++i) {
         params_[i].name = other.params_[i].name;
         if (auto p = interface_cast<ICloneable>(other.params_[i].value)) {
             params_[i].value = interface_pointer_cast<IAny>(p->GetClone());
